+++ ee/packages/pdf-worker/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @rocket.chat/pdf-worker
 
+## 0.0.4
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.2.9
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @rocket.chat/core-typings@6.2.7
+
 ## 0.0.2
 
 ### Patch Changes